Weight/confidence-based file replication
New objects' files are now stored in a staging directory (and never
deleted; bug), and maint will replicate them to peers according
to peer weight (storage per peer) and confidence.  Replica count is
affected by confidence: a replica target of 3 could be stored on three
peers of confidence 100, or 6 peers of confidence 50, or a combination
of them.

Currently replica count is hardcoded to 3.  If a replica count of 3
cannot be obtained due to lack of peers, this is currently silently
accepted.